saga_director/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
langgraph
langchain-core
langchain-openai
langchain-community
httpx
pytest
aiosqlite
numpy
shapely
langchain-ollama
langchain
python-dotenv
